pyproject.toml
[project]
name = "TelegramRag"
version = "0.1.0"
readme = "README.md"
description = "ITMO RAG"
authors = [
    { name = "Ygrick", email = "igor2.004@mail.ru" }
]
requires-python = ">=3.12"
dependencies = [
    "pdfplumber>=0.11.7",
    "PyMuPDF>=1.26.0",
    "pandas>=2.2.0",
    "beautifulsoup4>=4.13.3",
    "requests>=2.32.0",
    "python-telegram-bot>=22.2",
    "python-dotenv>=1.0.1",
    "langchain-text-splitters>=0.3.4",
    "langchain>=0.3.26",
    "langchain-openai>=0.3.27",
    "langchain-community>=0.3.26",
    "langchain-huggingface>=0.3.0",
    "sentence-transformers>=5.0.0",
    "faiss-cpu>=1.11.0",
    "rank-bm25>=0.2.1",
    "torch>=2.5.1",
    "jq>=1.9.1",
    "pymupdf4llm>=0.0.26",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 120

[[tool.uv.index]]
name = "pytorch-cu121"
url = "https://download.pytorch.org/whl/cu121"
explicit = true

[tool.uv.sources]
torch = [{ index = "pytorch-cu121" }]
